+++ slangpy/__init__.py
@@ -48,10 +48,8 @@ from . import bindings
 from . import builtin as internal_marshalls
 
 # Torch integration
-from .torchintegration import TORCH_ENABLED
-
-if TORCH_ENABLED:
-    from .torchintegration import TorchModule
+# Only import things that load torch lazily here!
+from .torchintegration import TorchModule
 
 # Debug options for call data gen
 from .core.calldata import set_dump_generated_shaders, set_dump_slang_intermediates