+++ slangpy/torchintegration/__init__.py
@@ -1,21 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-# pyright: reportUnusedImport=false
 
-TORCH_ENABLED = False
+# Only import things that load torch lazily here!
+from .torchmodule import TorchModule
 
-try:
-    import torch  # @IgnoreException
-    from .torchfunction import TorchFunction
-    from .torchmodule import TorchModule
-    from .torchstruct import TorchStruct
 
-    __all__ = [
-        "TorchFunction",
-        "TorchModule",
-        "TorchStruct",
-    ]
-
-    TORCH_ENABLED = True
-
-except ImportError:
-    pass
+__all__ = [
+    "TorchModule",
+]